arduino sada se PWM moze upaliti i ugasiti
git-svn-id: svn+ssh://login/fpgasvn/arduino@2483 6164e1d0-f123-df11-80f4-0019d13f85ef

+++ hardware/f32c/mips/cores/f32c/wiring_analog.c
@@ -41,7 +41,7 @@ void analogOutputInit( void )
 
 void analogWrite(uint32_t ulPin, uint32_t ulValue)
 {
-  int32_t pwm_channel;
+  int8_t pwm_channel;
   
   if(ulPin >= variant_pin_map_size)
     return;
